python-backend/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
PyJWT
flask-cors
requests
